+++ db/migrate/001_create_manga_table.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+Sequel.migration do
+  change do
+    create_table(:manga) do
+      primary_key :id
+      String :adapter, null: false
+      String :url, null: false
+      String :name, null: false
+    end
+  end
+end