+++ lib/mangdown/chapter.rb
@@ -3,26 +3,26 @@
 module Mangdown
   # Mangdown chapter object, which holds pages
   class Chapter
+    extend Forwardable
+
     include Equality
-    include Enumerable
 
-    attr_reader :uri, :pages, :name, :manga, :chapter
-    attr_accessor :adapter
+    attr_reader :chapter
 
-    def initialize(uri, name, manga, chapter = nil)
-      @name = name
-      @manga = manga
+    def_delegators :chapter, :url, :name, :number
+
+    alias uri url
+
+    def initialize(chapter)
       @chapter = chapter
-      @uri = Addressable::URI.escape(uri)
-      @pages = []
     end
 
-    def each(&block)
-      @pages.each(&block)
+    def pages
+      @pages ||= chapter.pages.map { |page| Mangdown.page(page) }
     end
 
-    def to_chapter
-      self
+    def manga
+      @manga ||= Mangdown.manga(chapter.manga)
     end
 
     def path
@@ -31,8 +31,8 @@ module Mangdown
     alias to_path path
 
     def setup_path(dir = nil)
-      dir ||= File.join(DOWNLOAD_DIR, manga)
-      path = File.join(dir, name)
+      dir ||= manga.path
+      path = Tools.file_join(dir, name)
       path = Tools.valid_path_name(path)
       @path = Tools.relative_or_absolute_path(path)
     end
@@ -41,8 +41,11 @@ module Mangdown
       CBZ.one(dir)
     end
 
+    # rubocop:disable Metrics/AbcSize
+    # rubocop:disable Metrics/CyclomaticComplexity
+    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Metrics/PerceivedComplexity
     def download_to(dir = nil, opts = { force_download: false })
-      pages = map(&:to_page)
       failed = []
       succeeded = []
       skipped = []
@@ -53,10 +56,13 @@ module Mangdown
         setup_download_dir!(dir)
       end
 
-      Tools.hydra_streaming(pages, adapter.hydra_opts) do |stage, page, data = nil|
+      Tools.hydra_streaming(
+        pages,
+        chapter.hydra_opts
+      ) do |stage, page, data = nil|
         case stage
         when :failed
-          failed << page
+          failed << [page, data]
         when :succeeded
           succeeded << page
         when :before
@@ -72,52 +78,16 @@ module Mangdown
 
       { failed: failed, succeeded: succeeded, skipped: skipped }
     end
-
-    def load_pages
-      return @pages if @pages.any?
-
-      fetch_each_page { |page| @pages << page }
-      @pages.sort_by!(&:name)
-    end
+    # rubocop:enable Metrics/AbcSize
+    # rubocop:enable Metrics/CyclomaticComplexity
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/PerceivedComplexity
 
     private
 
     def setup_download_dir!(dir)
       setup_path(dir)
       FileUtils.mkdir_p(to_path) unless Dir.exist?(to_path)
-    end
-
-    def fetch_each_page
-      pages = build_page_hashes
-      page_data = Hash.new { |h, k| h[k] = +'' }
-      Tools.hydra_streaming(pages, adapter.hydra_opts) do |status, page, data = nil|
-        case status
-        when :before
-          true
-        when :body
-          page_data[page] << data
-        when :complete
-          page = get_page(page.uri, Nokogiri::HTML(page_data[page]))
-          yield(page)
-        end
-      end
-    end
-
-    def build_page_hashes
-      adapter.page_list.map do |page|
-        page[:chapter] = name
-        page[:manga] = manga
-        MDHash.new(page)
-      end
-    end
-
-    def get_page(uri, doc)
-      adapter = Mangdown.adapter!(uri, nil, doc)
-      page = adapter.page
-      page[:chapter] = name
-      page[:manga] = manga
-
-      MDHash.new(page)
     end
   end
 end