+++ test/lib/mangdown/support/logging_test.rb
@@ -1,10 +1,11 @@
+# frozen_string_literal: true
+
 require 'stringio'
 require 'test_helper'
 
 module Mangdown
   class LoggingTest < Minitest::Test
-
-    class LoggingEnabled 
+    class LoggingEnabled
       include Logging
 
       def log(message, level)
@@ -31,6 +32,7 @@ module Mangdown
       Mangdown.configure_logger(file: STDOUT)
     end
 
+    # rubocop:disable Metrics/AbcSize
     def test_logger
       @instance.log('hello world', :debug)
       @instance.log('hello world', :fatal)
@@ -40,6 +42,7 @@ module Mangdown
       assert_equal 1, @io.string.scan(/FATAL/).length
       assert_equal 2, @io.string.scan(/hello world/).length
     end
+    # rubocop:enable Metrics/AbcSize
 
     def test_configure_logger
       assert_equal @io, @instance.device