+++ test/lib/mangdown/support/tools_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 require 'webmock/minitest'
 
@@ -25,15 +27,15 @@ module Mangdown
       url_with_path = 'http://www.anything.com/path/to/something'
       url_with_slash = 'http://www.anything.com/'
       url = 'http://www.anything.com'
-      
+
       assert_equal url, Tools.get_root(url)
       assert_equal url, Tools.get_root(url_with_slash)
       assert_equal url, Tools.get_root(url_with_path)
     end
 
     def test_relative_or_absolute_path
-      absolute = %w(/root to this path)
-      relative = %w(relative path from here)
+      absolute = %w[/root to this path]
+      relative = %w[relative path from here]
 
       absolute_path = Tools.relative_or_absolute_path(*absolute)
       relative_path = Tools.relative_or_absolute_path(*relative)
@@ -42,6 +44,8 @@ module Mangdown
       assert_equal Dir.pwd + '/relative/path/from/here', relative_path.to_s
     end
 
+    # rubocop:disable Metrics/AbcSize
+    # rubocop:disable Metrics/MethodLength
     def test_valid_path_name
       assert_equal 'Name 00001.ext', Tools.valid_path_name('Name 001.ext')
       assert_equal 'Name 00001', Tools.valid_path_name('Name 001')
@@ -57,6 +61,8 @@ module Mangdown
         '/path/to/Name 00001', Tools.valid_path_name('/path/to/Name 001')
       )
     end
+    # rubocop:enable Metrics/AbcSize
+    # rubocop:enable Metrics/MethodLength
 
     def test_file_type
       image_dir = Pathname.new(
@@ -68,8 +74,11 @@ module Mangdown
       assert_equal 'gif', Tools.image_extension(image_dir.join('naruto.gif'))
     end
 
+    # rubocop:disable Metrics/AbcSize
+    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Metrics/CyclomaticComplexity
     def test_hydra_streaming
-      objects = 3.times.map do |i|
+      objects = Array.new(3) do |i|
         url = "http://www.anything-#{i}.com/"
         stub_request(:get, url).to_return(
           body: "<html><body>Body-#{i}</body>"
@@ -84,13 +93,14 @@ module Mangdown
       bodies = []
       fails = []
       complete = []
-      
-      Tools.hydra_streaming(objects) do |status, object, data=nil|
+
+      Tools.hydra_streaming(objects) do |status, object, data = nil|
         case status
         when :before
           return true unless objects.index(object) == 1
         when :succeeded
           # Do nothing
+          nil
         when :failed
           fails << object
         when :complete
@@ -105,5 +115,8 @@ module Mangdown
       refute_includes complete, objects.delete(1)
       assert_equal objects, complete
     end
+    # rubocop:enable Metrics/AbcSize
+    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/CyclomaticComplexity
   end
 end